+++ gha_utils/mailmap.py
@@ -18,37 +18,78 @@ from __future__ import annotations
 
 import logging
 import sys
+from dataclasses import dataclass
 from functools import cached_property
 from subprocess import run
-from textwrap import dedent
+
+
+@dataclass(order=True, frozen=True)
+class Record:
+    """A mailmap identity mapping entry."""
+
+    # Mapping is define as the first field so we have natural sorting, whatever the value of the pre_comment is.
+    mapping: str
+    pre_comment: str
+
+    def __str__(self) -> str:
+        """Render the record with pre-comments first, followed by the identity mapping."""
+        lines = []
+        if self.pre_comment:
+            lines.append(self.pre_comment)
+        if self.mapping:
+            lines.append(self.mapping)
+        return "\n".join(lines)
 
 
 class Mailmap:
-    """Helpers to manipulate ``.mailmap`` file.
+    """Helpers to manipulate ``.mailmap`` files.
 
-    The ``.mailmap`` files expected to be found in the root of repository.
+    ``.mailmap`` `file format is documented on Git website
+    <https://git-scm.com/docs/gitmailmap>`_.
     """
 
-    def __init__(self, initial_mailmap: str | None = None) -> None:
-        if not initial_mailmap:
-            # Initialize empty .mailmap with pointers to reference documentation.
-            self.content = dedent(
-                """\
-                # Format is:
-                #   Preferred Name <preferred e-mail>  Other Name <other e-mail>
-                #
-                # Reference: https://git-scm.com/docs/git-blame#_mapping_authors
-                """,
-            )
-        else:
-            self.content = initial_mailmap
-        logging.debug(f"Initial content set to:\n{self.content}")
+    # Use a set to naturally deduplicate identical records.
+    records: set[Record] = set()
+
+    def parse(self, content: str):
+        """Parse mailmap content and add it to the current list of records.
+
+        Each non-empty, non-comment line is considered a mapping entry.
+
+        The preceeding lines of a maping entry are kept attached to it as pre-comments,
+        so the layout will be preserved on rendering, during which records are sorted.
+        """
+        logging.debug(f"Parsing:\n{content}")
+        preceeding_lines = []
+        for line in content.splitlines():
+            # Comment lines are added as-is.
+            if line.startswith("#"):
+                preceeding_lines.append(line)
+            # Blank lines are added as-is.
+            elif not line.strip():
+                preceeding_lines.append(line)
+            # Mapping entry, which mark the end of a block, so add it to the list mailmap records.
+            else:
+                # TODO: Validates and plit mapping entry here.
+                record = Record(pre_comment="\n".join(preceeding_lines), mapping=line)
+                logging.debug(record)
+                preceeding_lines = []
+                self.records.add(record)
+
+    def find(self, identity: str) -> bool:
+        """Returns ``True`` if the provided identity matched any record."""
+        for record in self.records:
+            # Identity matching is case insensitive: https://git-scm.com/docs/gitmailmap#_syntax
+            if identity.lower() in record.mapping.lower():
+                return True
+        return False
 
     @cached_property
     def git_contributors(self) -> set[str]:
         """Returns the set of all constributors found in the Git commit history.
 
-        No normalization happens: all variations of authors and committers strings attached to all commits are considered.
+        No normalization happens: all variations of authors and committers strings
+        attached to all commits are considered.
 
         For format output syntax, see:
         https://git-scm.com/docs/pretty-formats#Documentation/pretty-formats.txt-emaNem
@@ -72,30 +113,23 @@ class Mailmap:
         )
         return contributors
 
-    def updated_map(self):
-        """Add all missing contributors from commit history to mailmap.
+    def update_from_git(self) -> None:
+        """Add to internal records all missing contributors found in commit history.
 
         This method will refrain from adding contributors already registered as aliases.
         """
-        # Extract comments in .mailmap header and keep mapping lines.
-        header_comments = []
-        mappings = set()
-        for line in self.content.splitlines():
-            if line.startswith("#"):
-                header_comments.append(line)
-            elif line.strip():
-                mappings.add(line)
-
-        # Add all missing contributors to the mail mapping.
         for contributor in self.git_contributors:
-            if contributor not in self.content:
-                logging.debug(f"{contributor!r} not found in original content, add it.")
-                mappings.add(contributor)
+            if not self.find(contributor):
+                logging.info(f"Add new identity {contributor}")
+                self.records.add(contributor)
+            else:
+                logging.debug(f"Ignore existing identity {contributor}")
 
-        # Render content in .mailmap format.
+    def render(self) -> str:
+        """Render internal records in Mailmap format."""
         return (
-            "\n".join(header_comments)
-            + "\n\n"
-            + "\n".join(sorted(mappings, key=str.casefold))
+            "\n".join(
+                map(str, sorted(self.records, key=lambda r: r.mapping.casefold()))
+            )
             + "\n"
         )