+++ changelog.md
@@ -7,7 +7,8 @@
 
 - Rename `gha-utils mailmap` command to `gha-utils mailmap-sync`.
 - Add new `--create-if-missing`/`--skip-if-missing` option to `gha-utils mailmap-sync` command.
-- Do not create `.mailmap` from scratch in workflows: only update existing ones.
+- Keep comments attached to their mapping when re-sorting `.mailmap` files.
+- Do not keep adding header metadata on `.mailmap` updates.
 
 ## [4.1.4 (2024-07-02)](https://github.com/kdeldycke/workflows/compare/v4.1.3...v4.1.4)